fix: after killing child processes, kill parent process by sending a SIGTERM first
* fix: after killing child processes, kill parent process by sending a SIGTERM first

* chore: ensure only living processes are killed

---------

+++ src/main/java/com/aws/greengrass/util/platforms/unix/UnixExec.java
@@ -123,7 +123,7 @@ public class UnixExec extends Exec {
                 for (PidProcess pp : pidProcesses) {
                     pp.waitFor(gracefulShutdownTimeout.getSeconds(), TimeUnit.SECONDS);
                 }
-                if (pidProcesses.stream().anyMatch(pidProcess -> {
+                boolean isAnyProcessAlive = pidProcesses.stream().anyMatch(pidProcess -> {
                     try {
                         return pidProcess.isAlive();
                     } catch (IOException ignored) {
@@ -131,12 +131,13 @@ public class UnixExec extends Exec {
                         Thread.currentThread().interrupt();
                     }
                     return false;
-                })) {
+                });
+                if (isAnyProcessAlive) {
                     logger.atWarn()
                             .log("Command {} did not respond to interruption within timeout. Going to kill it now",
                                     this);
+                    platformInstance.killProcessAndChildren(p, true, pids, userDecorator);
                 }
-                platformInstance.killProcessAndChildren(p, true, pids, userDecorator);
                 if (!p.waitFor(5, TimeUnit.SECONDS) && !isClosed.get()) {
                     throw new IOException("Could not stop " + this);
                 }
